backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.4.2
antlr4-python3-runtime==4.9.2
